♻️  Refact sidebar filesystem structure
