+++ frontend/app/chat/components/ChatsList/components/SidebarFooter/components/BrainManagementButton.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { FaBrain } from "react-icons/fa";
 
-import { sidebarLinkStyle } from "@/app/chat/components/ChatsList/components/ChatsListItem/styles/SidebarLinkStyle";
+import { sidebarLinkStyle } from "@/app/chat/components/ChatsList/components/SidebarFooter/styles/SidebarLinkStyle";
 import { useBrainContext } from "@/lib/context/BrainProvider/hooks/useBrainContext";
 
 export const BrainManagementButton = (): JSX.Element => {
