+++ frontend/app/chat/components/ChatsList/components/SidebarFooter/hooks/useGravatar.ts
