+++ frontend/app/chat/components/ChatsList/components/SidebarHeader.tsx
