+++ frontend/app/chat/components/ChatsList/index.tsx
@@ -4,40 +4,18 @@
 import { motion, MotionConfig } from "framer-motion";
 import { MdChevronRight } from "react-icons/md";
 
-import { useChatsContext } from "@/lib/context/ChatsProvider/hooks/useChatsContext";
 import { cn } from "@/lib/utils";
 
-import { ChatsListItem } from "./components/ChatsListItem";
-import { SidebarActions } from "./components/ChatsListItem/components/SidebarActions";
-import { SidebarHeader } from "./components/ChatsListItem/components/SidebarHeader";
+import { ChatHistory } from "./components/ChatHistory";
 import { NewChatButton } from "./components/NewChatButton";
+import { SidebarActions } from "./components/SidebarFooter/SidebarActions";
+import { SidebarHeader } from "./components/SidebarHeader";
 import { useChatNotificationsSync } from "./hooks/useChatNotificationsSync";
 import { useChatsList } from "./hooks/useChatsList";
-import {
-  isToday,
-  isWithinLast30Days,
-  isWithinLast7Days,
-  isYesterday,
-} from "./utils";
 
 export const ChatsList = (): JSX.Element => {
-  const { allChats } = useChatsContext();
-
   const { open, setOpen } = useChatsList();
   useChatNotificationsSync();
-
-  const todayChats = allChats.filter((chat) =>
-    isToday(new Date(chat.creation_time))
-  );
-  const yesterdayChats = allChats.filter((chat) =>
-    isYesterday(new Date(chat.creation_time))
-  );
-  const last7DaysChats = allChats.filter((chat) =>
-    isWithinLast7Days(new Date(chat.creation_time))
-  );
-  const last30DaysChats = allChats.filter((chat) =>
-    isWithinLast30Days(new Date(chat.creation_time))
-  );
 
   return (
     <MotionConfig transition={{ mass: 1, damping: 10 }}>
@@ -70,46 +48,7 @@ export const ChatsList = (): JSX.Element => {
             <div className="pt-2">
               <NewChatButton />
             </div>
-            <div
-              data-testid="chats-list-items"
-              className="flex-1 overflow-auto scrollbar h-full"
-            >
-              {todayChats.length > 0 && (
-                <div className="bg-gray-100 text-black rounded-md px-3 py-1 mt-2">
-                  Today
-                </div>
-              )}
-              {todayChats.map((chat) => (
-                <ChatsListItem key={chat.chat_id} chat={chat} />
-              ))}
-
-              {yesterdayChats.length > 0 && (
-                <div className="bg-gray-100 text-black rounded-md px-3 py-1 mt-2">
-                  Yesterday
-                </div>
-              )}
-              {yesterdayChats.map((chat) => (
-                <ChatsListItem key={chat.chat_id} chat={chat} />
-              ))}
-
-              {last7DaysChats.length > 0 && (
-                <div className="bg-gray-100 text-black rounded-md px-3 py-1 mt-2">
-                  Previous 7 Days
-                </div>
-              )}
-              {last7DaysChats.map((chat) => (
-                <ChatsListItem key={chat.chat_id} chat={chat} />
-              ))}
-
-              {last30DaysChats.length > 0 && (
-                <div className="bg-gray-100 text-black rounded-md px-3 py-1 mt-2">
-                  Previous 30 Days
-                </div>
-              )}
-              {last30DaysChats.map((chat) => (
-                <ChatsListItem key={chat.chat_id} chat={chat} />
-              ))}
-            </div>
+            <ChatHistory />
             <SidebarActions />
           </div>
         </motion.div>
